requirements.txt
streamlit
langchain
langchain-core
langchain-community
openai
python-dotenv
markdown
matplotlib
calplot
pillow
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
